// src/app/pages/blog/models/blog.model.ts
export interface BlogModel {
    ID: number;
    id: number;
    files: string[];
    title: string;
    title_url: string;
    resume: string;
    content: string;
    keyword: string[];
    signature: string;
    created: Date; 
    views: number;
    CreatedAt: Date; 
    UpdatedAt: Date; 
    created_at: Date; 
    updated_at: Date; 
}